Verificacion de eliminar a estudiante con materia asociada completo

+++ backend/controllers/studentsController.php
@@ -11,6 +11,8 @@
 
 require_once("./repositories/students.php");
 
+require_once("./repositories/studentsSubjects.php");
+
 // Para GET (usamos la variable superglobal $_GET):
 //https://www.php.net/manual/es/language.variables.superglobals.php
 function handleGet($conn) 
@@ -77,16 +79,28 @@ function handlePut($conn)
 function handleDelete($conn) 
 {
     $input = json_decode(file_get_contents("php://input"), true);
+    
+    $hayMateriaAsociada = getSubjectsByStudent($conn, $input['id']); //verifico si hay una materia asociada al estudiante
 
-    $result = deleteStudent($conn, $input['id']);
-    if ($result['deleted'] > 0) 
-    {
-        echo json_encode(["message" => "Eliminado correctamente"]);
-    } 
-    else 
-    {
-        http_response_code(500);
-        echo json_encode(["error" => "No se pudo eliminar"]);
+    error_log("Datos recibidos: " . json_encode($hayMateriaAsociada));
+    if(empty($hayMateriaAsociada)){ //si no la hay se borra normal
+        $result = deleteStudent($conn, $input['id']);
+        if ($result['deleted'] > 0) 
+        {
+            echo json_encode(["message" => "Eliminado correctamente"]);
+        } 
+        else 
+        {
+            http_response_code(500);
+            echo json_encode(["error" => "No se pudo eliminar"]);
+        }
+    } else {//si la hay agarro un nombre (Podria haber mas de 1, me quedo con el primero)
+        $nombrePrimerMat = $hayMateriaAsociada[0]['name'];
+
+        http_response_code(202); //Tomo como valida la peticion
+
+        echo json_encode(["error" => "No se puede eliminar estudiantes con materias",
+                          "materia" => $nombrePrimerMat]);// devuelvo el nombre de la materia
     }
 }
 ?>
